Rename functions and move to own file
